Basic lookup server complete

+++ cs/ENFLookupServer/ENFLookup/FreqDbMetaData.cs
@@ -3,7 +3,7 @@ namespace ENFLookup;
 public class FreqDbMetaData
 {
     public string GridId { get; set; }
-    public int StartDate { get; set; }
-    public int EndDate { get; set; }
+    public long StartDate { get; set; }
+    public long EndDate { get; set; }
     public int BaseFrequency { get; set; }
 }
